chore(app): rename to settings page
