+++ app/lib/common/models/userdata.dart
@@ -1,7 +1,7 @@
 import 'package:hive/hive.dart';
 
-import '../../profile/models/hive/cpic_lookup_response.dart';
-import '../../profile/models/hive/diplotype.dart';
+import '../../settings/models/hive/cpic_lookup_response.dart';
+import '../../settings/models/hive/diplotype.dart';
 
 part 'userdata.g.dart';
