+++ app/lib/settings/models/hive/diplotype.dart
