+++ app/lib/settings/pages/settings.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class SettingsPage extends StatelessWidget {
+  const SettingsPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Settings'),
+    );
+  }
+}
